requirements.txt
streamlit
openai
faiss-cpu
langchain
pypdf
python-dotenv
